Move admin user handlers to handler package
