+++ server/routes/system.go
@@ -2,7 +2,6 @@ package routes
 
 import (
 	"STUOJ/server/handler"
-	handler_admin "STUOJ/server/handler-admin"
 	"STUOJ/server/middlewares"
 	"github.com/gin-gonic/gin"
 )
@@ -12,8 +11,6 @@ func InitSystemRoute(ginServer *gin.Engine) {
 	{
 		// 使用中间件
 		rootPrivateRoute.Use(middlewares.TokenAuthRoot())
-
-		rootPrivateRoute.PUT("/user/role", handler_admin.AdminUserModifyRole)
 
 		rootPrivateRoute.GET("/config", handler.ConfigList)
 	}
