+++ server/routes/testcase.go
@@ -2,7 +2,6 @@ package routes
 
 import (
 	"STUOJ/server/handler"
-	"STUOJ/server/handler-admin"
 	"STUOJ/server/middlewares"
 
 	"github.com/gin-gonic/gin"
@@ -14,17 +13,10 @@ func InitTestcaseRoute(ginServer *gin.Engine) {
 		// 使用中间件
 		testcaseEditorRoute.Use(middlewares.TokenAuthAdmin())
 
-		{
-			testcaseEditorRoute.GET("/user", handler_admin.UserList)
-			testcaseEditorRoute.POST("/user", handler_admin.AdminUserAdd)
-			testcaseEditorRoute.DELETE("/user/:id", handler_admin.AdminUserRemove)
-		}
-		{
-			testcaseEditorRoute.GET("/:id", handler.TestcaseInfo)
-			testcaseEditorRoute.POST("/", handler.TestcaseAdd)
-			testcaseEditorRoute.PUT("/", handler.TestcaseModify)
-			testcaseEditorRoute.DELETE("/:id", handler.TestcaseRemove)
-			testcaseEditorRoute.POST("/datamake", handler.TestcaseDataMake)
-		}
+		testcaseEditorRoute.GET("/:id", handler.TestcaseInfo)
+		testcaseEditorRoute.POST("/", handler.TestcaseAdd)
+		testcaseEditorRoute.PUT("/", handler.TestcaseModify)
+		testcaseEditorRoute.DELETE("/:id", handler.TestcaseRemove)
+		testcaseEditorRoute.POST("/datamake", handler.TestcaseDataMake)
 	}
 }
